streamlit/requirements.txt
streamlit==1.29.0
openai==1.6.1
python-dotenv==1.0.0
boto3==1.34.11
rich==13.7.0
